preprocessing 4

+++ main/src/loss.py
@@ -12,9 +12,9 @@ def kl_divergence(model, source_output, target_output, percent_lambda):
 
 def CORAL(source_output, target_output, percent_lambda=0.5):
 
-    source_batch_size = tf.cast(tf.shape(source_output)[0], tf.float32)
-    target_batch_size = tf.cast(tf.shape(target_output)[0], tf.float32)
-    d = tf.cast(tf.shape(source_output)[1], tf.float32)
+    source_batch_size = tf.cast(tf.shape(source_output)[0], tf.float64)
+    target_batch_size = tf.cast(tf.shape(target_output)[0], tf.float64)
+    d = tf.cast(tf.shape(source_output)[1], tf.float64)
 
     # Source covariance
     xm = source_output - tf.reduce_mean(source_output, 0, keepdims=True)
